Enable Easier Project Releases
Uses https://github.com/rwjblue/create-rwjblue-release-it-setup

This creates a config that allows GitHub/NPM releases to be managed automatically by running `npx release-it`, including automatic changelog generation. Instructions for how to setup for use locally are in RELEASE.md.

The actual changelog will be updated in full for the next release.